refactor(core): extract pod/lifecycle.py — pure code move
Track A Sprint 1 Step 1 (pod_lifecycle). Moves start_pod and stop_pod
out of pod/__init__.py into pod/lifecycle.py. Public surface unchanged
('from winpodx.core.pod import start_pod' still works) via re-export.

test_backend.py: monkeypatches that target winpodx.core.pod.get_backend
to intercept start_pod's backend factory now point at the function's
actual module (winpodx.core.pod.lifecycle.get_backend), since that's
where the resolution happens at call time.
